limit stack traces to macos for now to keep tests compiling
